Changed color game spirte to buckets
Added the use of buckets, also made a new dir called PlayerSprite to hold the sprites for the player, will need to re-do sprite orders and mechanics to be less arab.

Now, the code changes the direction of "some" sprites, and some not, thus, we dont have ONE rule for handling sprites and it's bad.